# pkg/tts/handlers/openapi_types.gen.go
// Package handlers provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.4.1 DO NOT EDIT.
package handlers

// Error defines model for Error.
type Error struct {
	Message string `json:"message"`
	Slug    string `json:"slug"`
}

// TtsRequest defines model for TtsRequest.
type TtsRequest struct {
	// Lang Language code for the speech (e.g., "en" for English, "id" for Indonesian).
	Lang *string `json:"lang,omitempty"`
	Text string  `json:"text"`
}

// TtsResponse defines model for TtsResponse.
type TtsResponse struct {
	Url string `json:"url"`
}

// TextToSpeechJSONRequestBody defines body for TextToSpeech for application/json ContentType.
type TextToSpeechJSONRequestBody = TtsRequest

// ReadTextToSpeechJSONRequestBody defines body for ReadTextToSpeech for application/json ContentType.
type ReadTextToSpeechJSONRequestBody = TtsRequest